Remove some compiler warnings

+++ src/mesher_interface.h
@@ -6,7 +6,7 @@
 namespace mesher_interface {
 	void initialize();
 
-	// Loads STEP file into GMSH environment.
+	// Supports STEP file
 	int import_model(const std::string& filename);
 
 	void mesh_model(int mesh_size_min, int mesh_size_max);
@@ -14,8 +14,8 @@ namespace mesher_interface {
 	void view_model();
 
 	box get_bounding_box();
-	box get_bounding_box(int dim, size_t tag);
-	std::vector<box> get_bounding_box(int dim, std::vector<size_t> tags);
+	box get_bounding_box(int dim, int tag);
+	std::vector<box> get_bounding_box(int dim, std::vector<int> tags);
 
 	int add_box(box b);
